ai_recruitment_system/requirements.txt
moya-ai>=0.1.0
spacy>=3.0.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
numpy>=1.20.0

# After installing these packages, run:
# python -m spacy download en_core_web_sm
